requirements.txt
joblib==1.4.2
numpy==1.26.4
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
paddlepaddle-gpu==2.6.1
paddleocr==2.8.1
Pillow==9.4.0
pyvi==0.1.1
scikit_image==0.24.0
scikit_learn==1.5.1
torch==2.3.1+cu121
torchvision==0.18.1+cu121
vietocr==0.3.5